explore rhai scripting in sdk
