+++ craft_sdk/Cargo.toml
@@ -15,6 +15,7 @@ plonky2 = { workspace = true }
 env_logger = { workspace = true }
 time = "0.3.46"
 tinytemplate = "1.2.1"
+rhai = "1.24.0"
 pod2utils = { path = "../pod2utils" }
 txlib = { path = "../txlib" }
 vdfpod = { path = "../intro_pods/vdfpod" }
